backend/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
django-cors-headers==3.13.0
